workstation/requirements.txt
customtkinter
Pillow
paho-mqtt
bidict
pyinstaller
pygame
matplotlib
pandas
winshell
pypiwin32